feat: implement user and student existence checks during addition

+++ server/usercontrollers/addUsers.js
@@ -2,6 +2,11 @@ const fs = require("fs");
 const addUser = (req, res) => {
   const data = JSON.parse(fs.readFileSync("Users.json", "utf-8"));
   const User = req.body;
+  const UserId = data.map((user) => user.UserId);
+  const UserIndex = UserId.indexOf(User.UserId);
+
+  if (UserIndex > -1) throw "User Already Exists";
+
   data.push(User);
 
   fs.writeFileSync("Users.json", JSON.stringify(data, null, 2));
